Unify bump scripts and rename ValidateInput to PrepareInput
Collapse toml-test-bump-pin.ts and yaml-test-bump-pin.ts into a single
suite-bump-pin.ts that accepts the target file as its first argument.
Rename ValidateInput to PrepareInput since it both strips the UTF-8 BOM
(mutating FIndex) and validates control characters.
