+++ .github/workflows/toml-test-bump.yml
@@ -32,7 +32,7 @@ jobs:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Update pin in suite runner
-        run: bun scripts/toml-test-bump-pin.ts ${{ steps.latest.outputs.sha }}
+        run: bun scripts/suite-bump-pin.ts scripts/run_toml_test_suite.py ${{ steps.latest.outputs.sha }}
 
       - id: changed
         run: |
